// generated/src/app/api-models/smhostauth-s.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SmhostauthS {
  namedobjectIdSequenceid: number;
  namedobjectIdObjecttype: number;
  hostauthConsoleuser: string;
  hostauthOsusername: string;
  hostauthIsosusernameset: number;
  hostauthOspassword: string;
  hostauthIsospasswordset: number;
}